+++ src/main/java/at/porscheinformatik/seleniumcomponents/SeleniumComponent.java
@@ -2,6 +2,7 @@ package at.porscheinformatik.seleniumcomponents;
 
 import java.util.Objects;
 
+import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebElement;
 
 /**
@@ -32,11 +33,14 @@ public interface SeleniumComponent extends WebElementContainer
     {
         SeleniumComponent parent = parent();
 
-        Objects.requireNonNull(parent, () -> String.format(
-            "The parent is null which means, that this component is root. Implement this method by supplying a valid environment"));
+        Objects
+            .requireNonNull(parent, () -> String
+                .format(
+                    "The parent is null which means, that this component is root. Implement this method by supplying a valid environment"));
 
-        return Objects.requireNonNull(parent.environment(),
-            () -> String.format("The environment of the component defined by %s is null", parent.getClass()));
+        return Objects
+            .requireNonNull(parent.environment(),
+                () -> String.format("The environment of the component defined by %s is null", parent.getClass()));
     }
 
     /**
@@ -46,6 +50,51 @@ public interface SeleniumComponent extends WebElementContainer
      * @return true if the component exists
      */
     boolean isReady();
+
+    /**
+     * Waits until the component becomes ready.
+     *
+     * @param timeoutInSeconds the timeout in seconds
+     */
+    default void waitUntilReady(double timeoutInSeconds)
+    {
+        SeleniumAsserts
+            .assertThatSoon(timeoutInSeconds, "Component becomes ready: " + describe(), () -> this,
+                SeleniumMatchers.isReady());
+    }
+
+    /**
+     * Returns true if a {@link WebElement} described by this component is visible. By default it checks, if the element
+     * is visible. This method has no timeout, it does not wait for the component to become existent. <br>
+     * <br>
+     * It it NO good idea, to check for invisibility. If the component is not visible, Selenium always waits for some
+     * time. This causes tests to run slowly and timeouts to fail.
+     *
+     * @return true if the component is visible
+     */
+    default boolean isVisible()
+    {
+        try
+        {
+            return SeleniumUtils.retryOnStale(() -> element().isDisplayed());
+        }
+        catch (NoSuchElementException e)
+        {
+            return false;
+        }
+    }
+
+    /**
+     * Waits until the component becomes ready.
+     *
+     * @param timeoutInSeconds the timeout in seconds
+     */
+    default void waitUntilVisible(double timeoutInSeconds)
+    {
+        SeleniumAsserts
+            .assertThatSoon(timeoutInSeconds, "Component becomes visible: " + describe(), () -> this,
+                SeleniumMatchers.isVisible());
+    }
 
     @Override
     String toString();