+++ src/main/java/at/porscheinformatik/seleniumcomponents/clarity/DeprecatedFormsClarityCheckboxComponent.java
@@ -6,7 +6,7 @@ package at.porscheinformatik.seleniumcomponents.clarity;
 import static at.porscheinformatik.seleniumcomponents.WebElementSelector.*;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
-import at.porscheinformatik.seleniumcomponents.SelectableSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.ActiveSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 import at.porscheinformatik.seleniumcomponents.component.CheckboxComponent;
@@ -17,7 +17,7 @@ import at.porscheinformatik.seleniumcomponents.component.HtmlComponent;
  * 
  * @author Daniel Furtlehner
  */
-public class DeprecatedFormsClarityCheckboxComponent extends AbstractSeleniumComponent implements SelectableSeleniumComponent
+public class DeprecatedFormsClarityCheckboxComponent extends AbstractSeleniumComponent implements ActiveSeleniumComponent
 {
     private final HtmlComponent label = new HtmlComponent(this, selectByTagName("label"));
     private final CheckboxComponent checkbox = new CheckboxComponent(this);