+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/ButtonComponent.java
@@ -6,16 +6,14 @@ package at.porscheinformatik.seleniumcomponents.component;
 import static at.porscheinformatik.seleniumcomponents.WebElementSelector.*;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
-import at.porscheinformatik.seleniumcomponents.ClickableSeleniumComponent;
-import at.porscheinformatik.seleniumcomponents.DeactivateableSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.ActiveSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 
 /**
  * @author Daniel Furtlehner
  */
-public class ButtonComponent extends AbstractSeleniumComponent
-    implements ClickableSeleniumComponent, DeactivateableSeleniumComponent
+public class ButtonComponent extends AbstractSeleniumComponent implements ActiveSeleniumComponent
 {
     public ButtonComponent(SeleniumComponent parent)
     {