+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/HtmlComponent.java
@@ -1,7 +1,7 @@
 package at.porscheinformatik.seleniumcomponents.component;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
-import at.porscheinformatik.seleniumcomponents.ClickableSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.ActiveSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumUtils;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
@@ -13,7 +13,7 @@ import at.porscheinformatik.seleniumcomponents.WebElementSelector;
  * @author ham
  * @author Daniel Furtlehner
  */
-public final class HtmlComponent extends AbstractSeleniumComponent implements ClickableSeleniumComponent
+public final class HtmlComponent extends AbstractSeleniumComponent implements ActiveSeleniumComponent
 {
 
     /**