+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/ImageComponent.java
@@ -6,14 +6,14 @@ package at.porscheinformatik.seleniumcomponents.component;
 import static at.porscheinformatik.seleniumcomponents.WebElementSelector.*;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
-import at.porscheinformatik.seleniumcomponents.ClickableSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.ActiveSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 
 /**
  * @author Daniel Furtlehner
  */
-public class ImageComponent extends AbstractSeleniumComponent implements ClickableSeleniumComponent
+public class ImageComponent extends AbstractSeleniumComponent implements ActiveSeleniumComponent
 {
 
     public ImageComponent(SeleniumComponent parent)