+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/OptionComponent.java
@@ -1,6 +1,7 @@
 package at.porscheinformatik.seleniumcomponents.component;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.ActiveSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SelectableSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
@@ -8,7 +9,7 @@ import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 /**
  * @author cet
  */
-public class OptionComponent extends AbstractSeleniumComponent implements SelectableSeleniumComponent
+public class OptionComponent extends AbstractSeleniumComponent implements ActiveSeleniumComponent
 {
     public OptionComponent(SeleniumComponent parent, WebElementSelector selector)
     {